+++ src/pymdm/__init__.py
@@ -7,7 +7,7 @@ and Windows/Intune platforms.
 """
 
 __title__ = "pymdm"
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 
 
 from .command_runner import CommandRunner